Make some space in frame benchmarking.
